fix test recipe

+++ recipe/test.php
@@ -55,22 +55,43 @@ task(
     }
 );
 
+task('deploy:prepare', function () {
+    // Check if shell is POSIX-compliant
+    try {
+        cd(''); // To run command as raw.
+        $result = run('echo $0')->toString();
+        if ($result == 'stdin: is not a tty') {
+            throw new \RuntimeException(
+                "Looks like ssh inside another ssh.\n".
+                "Help: http://goo.gl/gsdLt9"
+            );
+        }
+    } catch (\RuntimeException $e) {
+        $formatter = Deployer::get()->getHelper('formatter');
+
+        $errorMessage = [
+            "Shell on your server is not POSIX-compliant. Please change to sh, bash or similar.",
+            "Usually, you can change your shell to bash by running: chsh -s /bin/bash",
+        ];
+        write($formatter->formatBlock($errorMessage, 'error', true));
+
+        throw $e;
+    }
+});
+
 task(
-    'prepare',
+    'test:prepare',
     function () {
         run('if [ ! -d {{deploy_path}} ]; then mkdir -p {{deploy_path}}; fi');
 
         // Check for existing /current directory (not symlink)
-        $result = run(
-            'if [ ! -L {{deploy_path}}/current ] && [ -d {{deploy_path}}/current ]; then echo true; fi'
-        )->toBool();
+        $result = run('if [ ! -L {{deploy_path}}/current ] && [ -d {{deploy_path}}/current ]; then echo true; fi')->toBool();
         if ($result) {
-            throw new \RuntimeException(
-                'There already is a directory (not symlink) named "current" in '.get(
-                    'deploy_path'
-                ).'. Remove this directory so it can be replaced with a symlink for atomic deployments.'
-            );
+            throw new \RuntimeException('There already is a directory (not symlink) named "current" in ' . get('deploy_path') . '. Remove this directory so it can be replaced with a symlink for atomic deployments.');
         }
+
+        // Create metadata .dep dir.
+        run("cd {{deploy_path}} && if [ ! -d .dep ]; then mkdir .dep; fi");
 
         // Create releases dir.
         run("cd {{deploy_path}} && if [ ! -d releases ]; then mkdir releases; fi");
@@ -88,12 +109,16 @@ task(
                 }
                 $src = $releaseMasterPath.DIRECTORY_SEPARATOR.$name;
                 $dst = parse('{{deploy_path}}').DIRECTORY_SEPARATOR.$name;
-                FileHelper::copyFile($src, $dst);
+                if (FileHelper::fileExists($src)) {
+                    FileHelper::copyFile($src, $dst);
+                } else {
+                    writeln($src.' skipped');
+                }
             }
         }
     }
 );
-before('deploy:lock', 'prepare');
+before('deploy:lock', 'test:prepare');
 
 /**
  * Delete useless branches, which no in remote repository
